requirements.txt
fastapi
uvicorn
python-jose
python-multipart
passlib
bcrypt
python-dotenv
sqlalchemy
pydantic-settings